# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"User": {
				"properties": {
					"id": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"position": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"firstName",
					"lastName",
					"email",
					"password",
					"position"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_User.Exclude_keyofUser.id__": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"position": {
						"type": "string"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"email",
					"password",
					"position"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_User.id_": {
				"$ref": "#/components/schemas/Pick_User.Exclude_keyofUser.id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"CreateUser": {
				"$ref": "#/components/schemas/Omit_User.id_"
			},
			"SignInResponse": {
				"properties": {
					"user": {
						"properties": {
							"position": {
								"type": "string"
							},
							"email": {
								"type": "string"
							},
							"lastName": {
								"type": "string"
							},
							"firstName": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"position",
							"email",
							"lastName",
							"firstName",
							"id"
						],
						"type": "object"
					},
					"accessToken": {
						"type": "string"
					},
					"success": {
						"type": "boolean"
					}
				},
				"required": [
					"user",
					"accessToken",
					"success"
				],
				"type": "object"
			},
			"LoginUser": {
				"properties": {
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"password",
					"email"
				],
				"type": "object"
			},
			"IResetPasswordSuccess": {
				"properties": {
					"message": {
						"type": "string"
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ForgotPasswordRequest": {
				"properties": {
					"email": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object"
			},
			"ResetPasswordRequest": {
				"properties": {
					"confirm_password": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"confirm_password",
					"password"
				],
				"type": "object"
			},
			"prioritiesEnum": {
				"enum": [
					"Low",
					"Normal",
					"High"
				],
				"type": "string"
			},
			"Task": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"assignees": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					},
					"priority": {
						"$ref": "#/components/schemas/prioritiesEnum"
					},
					"fileAttachment": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"title",
					"description",
					"startDate",
					"endDate",
					"assignees",
					"priority",
					"fileAttachment"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_Task.Exclude_keyofTask.id__": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"assignees": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					},
					"priority": {
						"$ref": "#/components/schemas/prioritiesEnum"
					},
					"fileAttachment": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"description",
					"startDate",
					"endDate",
					"assignees",
					"priority",
					"fileAttachment"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_Task.id_": {
				"$ref": "#/components/schemas/Pick_Task.Exclude_keyofTask.id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"CreateTask": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Omit_Task.id_"
					},
					{
						"properties": {
							"assignees": {
								"items": {
									"type": "string"
								},
								"type": "array"
							}
						},
						"required": [
							"assignees"
						],
						"type": "object"
					}
				]
			},
			"UpdateTask": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Omit_Task.id_"
					},
					{
						"properties": {
							"assignees": {
								"items": {
									"type": "string"
								},
								"type": "array"
							}
						},
						"required": [
							"assignees"
						],
						"type": "object"
					}
				]
			},
			"Partial_Omit_User.id__": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"position": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"UpdateUser": {
				"$ref": "#/components/schemas/Partial_Omit_User.id__"
			},
			"Pick_User.Exclude_keyofUser.password__": {
				"properties": {},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_User.password_": {
				"$ref": "#/components/schemas/Pick_User.Exclude_keyofUser.password__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"changedPassordResponse": {
				"$ref": "#/components/schemas/Omit_User.password_"
			},
			"ChangePasswordRequest": {
				"properties": {
					"confirm_password": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"oldPassword": {
						"type": "string"
					}
				},
				"required": [
					"confirm_password",
					"password",
					"oldPassword"
				],
				"type": "object"
			}
		},
		"securitySchemes": {
			"jwtAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Bearer token to access these API endpoints"
			}
		}
	},
	"info": {
		"title": "users-management",
		"version": "1.0.0",
		"description": "Users and tasks management backend",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/auth/signup": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Successfully created.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUser"
							}
						}
					}
				}
			}
		},
		"/api/auth/signin": {
			"post": {
				"operationId": "LoginUser",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SignInResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginUser"
							}
						}
					}
				}
			}
		},
		"/api/auth/forgot-password": {
			"post": {
				"operationId": "ForgotPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResetPasswordSuccess"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ForgotPasswordRequest"
							}
						}
					}
				}
			}
		},
		"/api/auth/reset-password/{token}": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResetPasswordSuccess"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ResetPasswordRequest"
							}
						}
					}
				}
			}
		},
		"/api/tasks": {
			"post": {
				"operationId": "CreateTask",
				"responses": {
					"201": {
						"description": "Successfully created.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Task"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTask"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllTasks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Task"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/tasks/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Task"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Task"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTask"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteTask",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/users/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUser"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/users": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/User"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/users/change-password": {
			"patch": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/changedPassordResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChangePasswordRequest"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}